applications/home/api/v1/serializers.py
from rest_framework import serializers
from applications.home.models import Contact

class ContactSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contact
        fields = '__all__'