requirements.txt
asgiref==3.7.2
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
distlib==0.3.6
Django==4.2.2
django-appconf==1.0.5
django-cors-headers==4.2.0
django-environ==0.9.0
django-filter==22.1
django-formtools==2.4.1
django-imagekit==4.1.0
django-import-export==3.2.0
django-model-utils==4.3.1
django-oauth-toolkit==2.3.0
django-recaptcha==3.0.0
django-resized==1.0.2
django-storages==1.13.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.12.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupPy==1.14
MarkupSafe==2.1.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pbr==5.11.1
pilkit==2.0
Pillow==9.4.0
platformdirs==3.2.0
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sendgrid==6.10.0
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
stevedore==5.0.0
tablib==3.4.0
tzdata==2022.6
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.22.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
whitenoise==6.3.0
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0